fix: 멤버 여부 확인 find -> exist로 변환 후 테스트코드 추가, advice에 basePackage 적용
