+++ src/main/java/com/example/sinitto/member/exception/NotUniqueException.java
@@ -1,4 +1,4 @@
-package com.example.sinitto.auth.exception;
+package com.example.sinitto.member.exception;
 
 public class NotUniqueException extends RuntimeException {
     public NotUniqueException(String message) {
